+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Release History
 ===============
 
+dev
+---
+
+* Rename Repository abstract methods to be public (Ex. `_create_object` → `create`)
+* Add `delete_all()` method to Entity to support Repository cleanup
+
 0.0.10 (2019-04-05)
 -------------------