+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ dev
 
 * Rename Repository abstract methods to be public (Ex. `_create_object` → `create`)
 * Add `delete_all()` method to Entity to support Repository cleanup
+* Remove requirement for explicit Model definitions for Entities
+    * Move Model options into Entity `Meta` class
 
 0.0.10 (2019-04-05)
 -------------------