+++ CHANGELOG.rst
@@ -7,8 +7,9 @@ dev
 
 * Rename Repository abstract methods to be public (Ex. `_create_object` → `create`)
 * Add `delete_all()` method to Entity to support Repository cleanup
+* Add support for `raw` queries on Entity repositories
 * Remove requirement for explicit Model definitions for Entities
-    * Move Model options into Entity `Meta` class
+* Move Model options into Entity `Meta` class
 
 0.0.10 (2019-04-05)
 -------------------